Implement a simple RTP streamer
This replaces the previous GoRTP-based implementation because I'm
tired of spending time chasing down null pointer derefences. I'm not
sure is there a global context I'm abusing or what but adding a new
session while the stream is already going on proved to be quite a
headache, so much so that writing your own RTP streamer is simpler.